TST: skips in extension.test_categorical

+++ pandas/tests/extension/base/methods.py
@@ -394,7 +394,10 @@ class BaseMethodsTests(BaseExtensionTests):
 
     def test_searchsorted(self, data_for_sorting, as_series):
         b, c, a = data_for_sorting
-        arr = type(data_for_sorting)._from_sequence([a, b, c])
+        # Pass dtype so sorting order is not lost for Categorical
+        arr = type(data_for_sorting)._from_sequence(
+            [a, b, c], dtype=data_for_sorting.dtype
+        )
 
         if as_series:
             arr = pd.Series(arr)
